requirements.txt
datasets>=2.14.0
weaviate-client>=4.0.0
transformers>=4.35.0
torch>=2.0.0
tiktoken>=0.5.0
numpy>=1.24.0
tqdm>=4.66.0
pyarrow>=12.0.0
pandas>=2.0.0

